Split testing of vectors into two different files, one for constructors and basic operations and other one for properties
