+++ test/src/utils/vectors/vector_basic.test.cpp
@@ -458,289 +458,6 @@ TEST_F(OperationsByConstant, handlesDivisionWithConstant)
     ASSERT_EQ(vector_result_one[i], static_cast<int>(vector_one[i]/ kValue));
   }
 }
-struct Associativity : public::testing::Test
-{
-  const int dimension = 3;
-  Vector<int, 3> vector_one;
-  Vector<int, 3> vector_two;
-  Vector<int, 3> vector_three;
-  Vector<int, 3> vector_result_one;
-  Vector<int, 3> vector_result_two;
-  void SetUp()
-{
-    vector_one = Vector<int, 3>(createRandomArray());
-    vector_two = Vector<int, 3>(createRandomArray());
-    vector_three = Vector<int, 3>(createRandomArray());
-    vector_result_one = Vector<int, 3>();
-    vector_result_two = Vector<int, 3>();
-  }
-};
-
-TEST_F(Associativity, isAdditionAssociative)
-{
-  vector_result_one += vector_one;
-  vector_result_one += vector_two;
-  vector_result_one += vector_three;
-  vector_result_two += vector_two;
-  vector_result_two += vector_three;
-  vector_result_two += vector_one;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector_result_two[i], vector_result_one[i]);
-  }
-}
-TEST_F(Associativity, isAutoAdditionAssociative)
-{
-  vector_result_one = (vector_one+ vector_two) + vector_three;
-  vector_result_two = vector_one + (vector_two+vector_three);
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector_result_two[i], vector_result_one[i]);
-  }
-}
-TEST_F(Associativity, isMultiplicationAssociative)
-{
-  vector_result_one *= vector_one;
-  vector_result_one *= vector_two;
-  vector_result_one *= vector_three;
-  vector_result_two *= vector_two;
-  vector_result_two *= vector_three;
-  vector_result_two *= vector_one;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector_result_two[i], vector_result_one[i]);
-  }
-}
-TEST_F(Associativity, isAutoMultiplicationAssociative)
-{
-  vector_result_one = (vector_one * vector_two) * vector_three;
-  vector_result_two = vector_one * (vector_two * vector_three);
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector_result_two[i], vector_result_one[i]);
-  }
-}
-struct Commutativity : public ::testing::Test
-{
-  const int dimension = 3;
-  Vector<int, 3> vector_one;
-  Vector<int, 3> vector_two;
-  Vector<int, 3> vector_result_one;
-  Vector<int, 3> vector_result_two;
-  void SetUp()
-  {
-    vector_one = Vector<int, 3>(createRandomArray());
-    vector_two = Vector<int, 3>(createRandomArray());
-    vector_result_one = Vector<int, 3>();
-    vector_result_two = Vector<int, 3>();
-  }
-};
-TEST_F(Commutativity, isAdditionCommutative)
-{
-  vector_result_one += vector_one;
-  vector_result_one += vector_two;
-  vector_result_two += vector_two;
-  vector_result_two += vector_one;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector_result_two[i], vector_result_one[i]);
-  }
-}
-TEST_F(Commutativity, isSubstractionNotCommutative)
-{
-  vector_result_one += vector_one;
-  vector_result_one -= vector_two;
-  vector_result_two += vector_two;
-  vector_result_two -= vector_one;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector_result_two[i], -vector_result_one[i]);
-  }
-}
-TEST_F(Commutativity, isAutoAdditionCommutative)
-{
-  vector_result_one = vector_two + vector_one;
-  vector_result_two = vector_one + vector_two;
-  for (int i = 0;i < dimension;i++) {
-  ASSERT_EQ(vector_result_one[i], vector_result_two[i]);
-  }
-}
-TEST_F(Commutativity, isAutoSubstractionNotCommutative)
-{
-  vector_result_one = vector_two - vector_one;
-  vector_result_two = vector_one - vector_two;
-  for (int i = 0;i < dimension;i++) {
-  ASSERT_EQ(vector_result_one[i], -vector_result_two[i]);
-  }
-}
-TEST_F(Commutativity, isMultiplicationCommutative)
-{
-  vector_result_one *= vector_two;
-  vector_result_one *= vector_one;
-  vector_result_two *= vector_one;
-  vector_result_two *= vector_two;
-  for (int i = 0;i < dimension;i++) {
-  ASSERT_EQ(vector_result_one[i], vector_result_two[i]);
-  }
-}
-TEST_F(Commutativity, isAutoMultiplicationCommutative)
-{
-  vector_result_one = vector_two * vector_one;
-  vector_result_two = vector_one * vector_two;
-  for (int i = 0;i < dimension;i++) {
-  ASSERT_EQ(vector_result_one[i], vector_result_two[i]);
-  }
-}
-
-struct IdentityOperations : public ::testing::Test
-{
-  std::array<int, 3> values = createRandomArray();
-  const int dimension = 3;
-  Vector<int, 3> identity_vector;
-  Vector<int, 3> vector;
-  void SetUp()
-  {
-    identity_vector = Vector<int, 3>();
-    vector = Vector<int, 3>(values);
-  }
-};
-
-TEST_F(IdentityOperations, handlesAdditionIdentity)
-{
-  vector+=identity_vector;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector[i], values[i]);
-  }
-}
-TEST_F(IdentityOperations, handlesSubstractionIdentities)
-{
-  vector -= identity_vector;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector[i], values[i]);
-  }
-  vector -= vector;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(vector[i], identity_vector[i]);
-  }
-}
-TEST_F(IdentityOperations, hadlesAutoAdditonIdentity)
-{
-  Vector<int, 3> output = vector + identity_vector;
-  for (int i = 0; i < dimension;i++) {
-    ASSERT_EQ(vector[i], output[i]);
-  }
-}
-TEST_F(IdentityOperations, handlesAutoSubstractionIdentities)
-{
-  Vector<int, 3> output = vector - identity_vector;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(output[i], values[i]);
-  }
-  output = vector - vector;
-  for (int i = 0;i < dimension;i++) {
-    ASSERT_EQ(output[i], identity_vector[i]);
-  }
-}
-TEST_F(IdentityOperations, handlesAutoMultiplicationIdentities)
-{
-  identity_vector = Vector<int, 3>(1);
-  Vector<int, 3> output = vector * identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(output[i], vector[i]);
-  }
-  identity_vector = Vector<int, 3>();
-  output = vector * identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(output[i], 0);
-  }
-}TEST_F(IdentityOperations, handlesMultiplicationIdentities)
-{
-  identity_vector = Vector<int, 3>(1);
-  vector *= identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(vector[i], values[i]);
-  }
-  identity_vector = Vector<int, 3>();
-  vector *= identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(vector[i], 0);
-  }
-}
-TEST_F(IdentityOperations, handlesChangeOfSignIdentity)
-{
-  identity_vector = -identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(identity_vector[i], 0);
-  }
-  Vector<int, 3> vector_two = -vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(identity_vector[i], vector_two[i] + vector[i]);
-  }
-}
-TEST_F(IdentityOperations, handlesAutoDivisionIdentities)
-{
-  vector = Vector<int, 3>(createRandomWithoutZeroesArray());
-  Vector<int, 3> output =  identity_vector / vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(output[i], identity_vector[i]);
-  }
-  identity_vector = Vector<int, 3>(1);
-  output =  vector/vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(output[i], identity_vector[i]);
-  }
-  output =  vector/identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(vector[i], output[i]);
-  }
-}
-TEST_F(IdentityOperations, handlesDivisionIdentities)
-{
-  vector = Vector<int, 3>(createRandomWithoutZeroesArray());
-  identity_vector /= vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(0, identity_vector[i]);
-  }
-  identity_vector = Vector<int, 3>(1);
-  vector /= vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(vector[i], identity_vector[i]);
-  }
-  std::array<int, 3>values = createRandomWithoutZeroesArray();
-  vector = Vector<int, 3>(values);
-  vector /=  identity_vector;
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(vector[i], values[i]);
-  }
-}
-struct EqualityOperation: public :: testing::Test
-{
-  const int dimension = 3;
-  std::array<int, 3> values;
-  Vector<int, 3> vector_one;
-  Vector<int, 3> vector_two;
-  void SetUp()
-  {
-    values = createRandomArray();
-    vector_one = Vector<int, 3>(values);
-    vector_two = Vector<int, 3>(values);
-  }
-};
-
-TEST_F(EqualityOperation, handlesEqualityIdentity)
-{
-  Vector<int, 3> identiy = Vector<int, 3>();
-  Vector<int, 3> vector_two = vector_one - identiy;
-  ASSERT_TRUE(vector_one == vector_two);
-  vector_two = vector_one - vector_one;
-  ASSERT_TRUE(vector_two == identiy);
-}
-TEST_F(EqualityOperation, handlesEquality)
-{
-  for (int i = 0; i < dimension; i++) {
-    ASSERT_EQ(vector_one[i], vector_two[i]);
-  }
-  ASSERT_TRUE(vector_one == vector_two);
-}
-TEST_F(EqualityOperation, handlesSymmetry)
-{
-  ASSERT_TRUE(vector_one == vector_two);
-  ASSERT_TRUE(vector_two == vector_one);
-}
 }
 }
 }
